Added method for sides
getSides returns an arraylist of Point3D arrays of length 3.

So a vertice will be a point3D object. Three vertices define a side. Thus each side is a Point3D[3] (ie a size 3 arrray of Point3D). 

sides is an arraylist of these.

There can be a little bit of optimisation if needed, currently L,P and T objects will have 24 sides, they can be optimised to get 22 sides